Breaking: use `renderSubtreeIntoContainer` for passing context (fixes #1)
