+++ .gitignore
@@ -25,7 +25,3 @@ coverage
 
 # Build directory
 /lib
-/react-redux-element-portal/lib
-/react-redux-element-portal/package.json
-/react-redux-element-portal/node_modules
-/test/helpers/ReduxElementPortal.js
