--- publish-packages.js
@@ -1,8 +1,0 @@
-import shell from 'shelljs';
-
-const cwd = process.cwd();
-
-shell.cd('react-redux-element-portal');
-shell.exec('npm publish', () => {
-  shell.cd(cwd);
-});
