--- react-redux-element-portal/README.md
@@ -1,5 +1,0 @@
-## React Redux Element Portal
-
-This is just `react-element-portal`, specifically bound to Redux, for convenience.
-
-See [react-element-portal](https://github.com/zapier/react-element-portal) for documentation.
