--- react-redux-element-portal/src/ElementPortal.js
@@ -1,6 +1,0 @@
-import { createElementPortal } from 'react-element-portal';
-import { Provider } from 'react-redux';
-
-const ElementPortal = createElementPortal(Provider, 'store');
-
-export default ElementPortal;
